Handle hidden folders better
The Canvas API returns a flat list of all folders in the course including all subfolders too.
But if a folder is locked/hidden, then that folder is not returned in the flat list but for some inexplicable reason, its subfolders are.
So assume that if we find a folder without a parent, then the parent is hidden from us, and so we ignore this folder too.